Initial implementation of wit storage
