+++ workitemtype.go
@@ -1,62 +1,35 @@
 package main
 
 import (
-	"fmt"
-	"strconv"
+	"log"
 
 	"github.com/almighty/almighty-core/app"
 	"github.com/almighty/almighty-core/models"
 	"github.com/goadesign/goa"
 )
 
-var wellKnown = map[string]*models.WorkItemType{
-	"1": &models.WorkItemType{
-		ID:   1,
-		Name: "system.workitem",
-		Fields: map[string]models.FieldDefinition{
-			"system.owner": models.FieldDefinition{Type: models.SimpleType{Kind: models.KindUser}, Required: true},
-			"system.state": models.FieldDefinition{Type: models.SimpleType{Kind: models.KindString}, Required: true},
-		}}}
-
 // WorkitemtypeController implements the workitemtype resource.
 type WorkitemtypeController struct {
 	*goa.Controller
+	witRepository models.WorkItemTypeRepository
 }
 
 // NewWorkitemtypeController creates a workitemtype controller.
-func NewWorkitemtypeController(service *goa.Service) *WorkitemtypeController {
-	return &WorkitemtypeController{Controller: service.NewController("WorkitemtypeController")}
+func NewWorkitemtypeController(service *goa.Service, witRepository models.WorkItemTypeRepository) *WorkitemtypeController {
+	return &WorkitemtypeController{Controller: service.NewController("WorkitemtypeController"), witRepository: witRepository}
 }
 
 // Show runs the show action.
 func (c *WorkitemtypeController) Show(ctx *app.ShowWorkitemtypeContext) error {
-	res, _ := loadTypeFromDB(ctx.ID)
-	if res != nil {
-		converted := convertTypeFromModels(*res)
-		return ctx.OK(&converted)
-	}
-	return ctx.NotFound()
-}
-
-func loadTypeFromDB(id string) (*models.WorkItemType, error) {
-	if wellKnown[id] == nil {
-		return nil, fmt.Errorf("Work item type not found: %s", id)
-	}
-	return wellKnown[id], nil
-}
-
-func convertTypeFromModels(t models.WorkItemType) app.WorkItemType {
-	var converted = app.WorkItemType{
-		ID:      strconv.FormatUint(t.ID, 10),
-		Name:    t.Name,
-		Version: 0,
-		Fields:  map[string]*app.FieldDefinition{},
-	}
-	for name, def := range t.Fields {
-		converted.Fields[name] = &app.FieldDefinition{
-			Required: def.Required,
-			Type:     def.Type,
+	res, err := c.witRepository.Load(ctx.Context, ctx.ID)
+	if err != nil {
+		switch err.(type) {
+		case models.NotFoundError:
+			log.Printf("not found, id=%s", ctx.ID)
+			return goa.ErrNotFound(err.Error())
+		default:
+			return err
 		}
 	}
-	return converted
+	return ctx.OK(res)
 }
